+++ src/components/combo_box/combo_box_options_list/index.ts
@@ -1,0 +1,6 @@
+export {
+  EuiComboBoxOptionsList,
+  EuiComboBoxOptionsListProps,
+} from './combo_box_options_list';
+export { EuiComboBoxOption, EuiComboBoxOptionProps } from './combo_box_option';
+export { EuiComboBoxTitle } from './combo_box_title';