+++ src/components/selectable/index.ts
@@ -1,4 +1,9 @@
-export { EuiSelectable } from './selectable';
-export { EuiSelectableList, EuiSelectableListItem } from './selectable_list';
+export { EuiSelectable, EuiSelectableProps } from './selectable';
+export {
+  EuiSelectableList,
+  EuiSelectableListItem,
+  EuiSelectableOptionsListProps,
+} from './selectable_list';
 export { EuiSelectableMessage } from './selectable_message';
+export { EuiSelectableOption } from './selectable_option';
 export { EuiSelectableSearch } from './selectable_search';